requirements.txt
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
coverage==5.5
deprecation==2.1.0
Django @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/django_1699555153998/work
eyed3==0.9.7
filetype==1.2.0
lxml==5.3.0
packaging==24.1
piexif==1.1.3
pillow==10.4.0
PyPDF2==3.0.1
pytaglib==3.0.0
python-docx==1.1.2
setuptools==72.1.0
sqlparse @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/sqlparse_1699544474746/work
toml==0.10.2
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
wheel==0.43.0
